Fix a failure of ndb_alter_table_online.test. Suppress a Windows compilation warning.

+++ vio/viosocket.c
@@ -372,7 +372,8 @@ my_bool vio_poll_read(Vio *vio,uint timeout)
   FD_ZERO(&errorfds);
   FD_SET(fd, &readfds);
   FD_SET(fd, &errorfds);
-  if ((res= select(fd, &readfds, NULL, &errorfds, &tm) <= 0))
+  /* The first argument is ignored on Windows, so a conversion to int is OK */
+  if ((res= select((int) fd, &readfds, NULL, &errorfds, &tm) <= 0))
   {
     DBUG_RETURN(res < 0 ? 0 : 1);
   }
